refactor: remove SocialAccount model and factory and move social account to a trait
