--- app/Models/SocialAccount.php
@@ -1,36 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\BelongsTo;
-
-class SocialAccount extends Model
-{
-    use HasFactory;
-
-    protected $fillable = [
-        'user_id',
-        'provider',
-        'provider_id',
-        'provider_token',
-        'provider_refresh_token',
-        'provider_token_expires_at',
-        'provider_avatar_url',
-        'last_login_at',
-    ];
-
-    protected function casts(): array
-    {
-        return [
-            'provider_token_expires_at' => 'datetime',
-            'last_login_at' => 'datetime',
-        ];
-    }
-
-    public function user(): BelongsTo
-    {
-        return $this->belongsTo(User::class);
-    }
-}
