+++ app/Models/User.php
@@ -4,19 +4,23 @@ namespace App\Models;
 
 // use Illuminate\Contracts\Auth\MustVerifyEmail;
 use App\Enums\Role;
-use Database\Factories\UserFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Spatie\MediaLibrary\HasMedia;
 use Spatie\MediaLibrary\InteractsWithMedia;
 use Spatie\Permission\Traits\HasRoles;
 
+//TODO: add to auth module install doc
+use Modules\Auth\Traits\HasSocialAccounts;
+
 class User extends Authenticatable implements HasMedia
 {
-    /** @use HasFactory<UserFactory> */
-    use HasFactory, HasRoles, InteractsWithMedia, Notifiable;
+    use HasFactory,
+        HasRoles,
+        InteractsWithMedia,
+        Notifiable,
+        HasSocialAccounts;
 
     /**
      * The attributes that are mass assignable.
@@ -72,11 +76,6 @@ class User extends Authenticatable implements HasMedia
             ->acceptsMimeTypes(['image/jpeg', 'image/png', 'image/gif', 'image/webp']);
     }
 
-    public function socialAccounts(): HasMany
-    {
-        return $this->hasMany(SocialAccount::class);
-    }
-
     /**
      * Get avatar with fallback to default
      */
@@ -97,16 +96,7 @@ class User extends Authenticatable implements HasMedia
         return asset('images/default-avatar.jpg');
     }
 
-    /**
-     * Get connected providers for profile display
-     */
-    public function getConnectedProvidersAttribute(): array
-    {
-        return $this->socialAccounts()
-            ->orderBy('last_login_at', 'desc')
-            ->get(['provider', 'last_login_at', 'provider_avatar_url'])
-            ->toArray();
-    }
+
 
     /**
      * Check if user is an administrator
