+++ modules/Auth/app/Models/SocialAccount.php
@@ -2,6 +2,7 @@
 
 namespace Modules\Auth\Models;
 
+use App\Models\User;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
